feat: enhance ESLint configuration and improve code quality
- Added configuration for script files to recognize `process` and `console` as readonly globals.
- Refactored various TypeScript files to improve type safety and readability, including updates to variable handling and method calls.
- Updated `Cargo.lock` to reflect the new version of the `voxelle-desktop` package.
- Made minor adjustments in Rust files for better error handling and code clarity.
- Improved performance in mesh generation and rendering processes by optimizing buffer handling and memory usage.

+++ src-tauri/src/generators/roof_gen.rs
@@ -230,7 +230,7 @@ fn roof_height_for_style(
     }
 
     let h = max_height.max(1) as f64;
-    let t = thickness.max(1) as i32;
+    let t = thickness.max(1);
 
     match style {
         // -- Flat styles --
